+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased (0.9.2)
+# v0.9.2
 * Log crf results, instead of printing, if stderr is not a terminal.
 * Wait for all child processes (ffmpeg etc) to finish before temp file cleanup and exit.